+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Merge "Places/Towns" and "Places/Settlements" layers into one "Populated
   places" layer. Scale point markers for this layer by population.
+* Add new "Geology/Mineral occurrences" layer from GEUS.
 
 
 # v3.0.0alpha4 (2023-07-21)